Refactor: /api/api.ts > /api/index.ts 파일명 변경
